[#61] [#174] Adds a terms and conditions page
This also covers the privacy policy

Note that I have written the page to cover two organisations
I have used the privacy policy from http://www.threesixtygiving.org/privacy/
and pasted that in.
The main text comes from the CoVE terms and conditions page, but with
some re-writes and some sections removed.
A link to this page is placed in the footer with some CSS to theme it
A test is added to check that this page exists and has a h1