frontend views Elastic 7 requires query lengths to be specified
Elastic Search requires that they queries used for funders_datatables
and funder_recipients_datatables have a length specified, rather than 0
which was used for infinite. This is not supported anymore so we set a
default length of 100k. This is overrieable by specifying a GET
parameter of length in the url.

Work towards resolving issue #510