frontend: views: Only allow searches with allowed default fields
We have seen cases on live server where people tried to search with odd
default fields: eg

default_field = *'A=0

In this case, the app tried to do the search then crashed when putting
together the view (in get_radio_items).

Instead, make sure we check that passed values are actually allowed.