js: Update jquery and bootstrap JS to more recent versions
This is because there's security issues with the existing versions.

Also make some changes for compatibility with jQuery 3.

jQuery is updated to the latest version (3), because the earlier major
versions don't have all security fixes backported.
https://security.snyk.io/package/npm/jquery

boostrap is updated to the latest minor version for the major version 3,
which we were using. This seems to have had all security fixes
backported.
https://security.snyk.io/package/npm/bootstrap