views: Add a results cache for the initial es query
This adds a cache that is useful for both template rendering and results from the aggregate API. Adding new data clears the whole cache here to ensure stale results are not kept. Settings are updated to use the file-based cache so that the cache is managed irrespective of the process.