dataload: import_to_elasticsearch: Remove field interdependency
In some geo areas some fields don't exist so avoid bailing out of the
whole process and handle each field's existence individually.

This is especially relevant to beneficiaryLocation due to the location
being able to be a geo area which is very large and therefore won't have
a single local authority.