Templates: New linkedin logo and footer changes
Had a go at resolving #1104 by doing simple find-and-replace, and
introducing a new LinkedIn logo. Used the "reversed white" variant to
match the other white logos we're using. Downloaded the logo from
fontawesome and manipulated it in inkscape to match the rough dimensions
of the other logos.

* https://brand.linkedin.com/content/brand/global/en_us/index/visual-identity/logo
* https://fontawesome.com/icons/linkedin?f=brands&s=solid

Unsure as to whether this will look ok as not sure what the CSS is doing
wrt to line height. Very happy for someone to come in and fix this if
it's borked.