import_to_elasticsearch: Add --limit to conveniently limit grants loaded
Often during testing it is useful to test with a smaller subset of the
data. This improves the use case of using a data package but not wanting
to load all the data. As the data is loaded in batches this will not be
an exact limit but to the nearest batch of data loaded.

usage:

--limit N